+++ src/traced/probes/probes_producer.cc
@@ -40,8 +40,8 @@
 namespace perfetto {
 namespace {
 
-uint64_t kInitialConnectionBackoffMs = 100;
-uint64_t kMaxConnectionBackoffMs = 30 * 1000;
+int kInitialConnectionBackoffMs = 100;
+int kMaxConnectionBackoffMs = 30 * 1000;
 constexpr char kFtraceSourceName[] = "linux.ftrace";
 constexpr char kProcessStatsSourceName[] = "linux.process_stats";
 constexpr char kInodeMapSourceName[] = "linux.inode_file_map";
